day-2/basic-drills/say-hello-p1/pyproject.toml
[project]
name = "say-hello-p1"
version = "0.1.2"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Pranitha", email = "pranithavanagala1998@gmai.com" }
]
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "testpypi"
url = "https://test.pypi.org/simple/"
publish-url = "https://test.pypi.org/legacy/"
explicit = true
